+++ mica-core/src/test/java/net/dreamlu/mica/test/bean/BeanTest1.java
@@ -1,7 +1,6 @@
 package net.dreamlu.mica.test.bean;
 
 import net.dreamlu.mica.core.utils.BeanUtil;
-import org.springframework.core.ResolvableType;
 
 import java.util.ArrayList;
 import java.util.List;